Bit of presenter reorganization
