+++ pakyow-presenter/lib/pakyow/extensions/router.rb
@@ -1,0 +1,10 @@
+require "pakyow/presenter/presentable"
+
+module Pakyow
+  class Router
+    prepend Presenter::Presentable
+    extend Presenter::Presentable::ClassMethods
+
+    def_delegators :controller, :render
+  end
+end
