+++ pakyow-presenter/lib/pakyow/presenter/container.rb
@@ -1,5 +1,6 @@
 module Pakyow
   module Presenter
+    # TODO: refactor to ContainerPresenter
     class Container < View
     end
   end
