+++ pakyow-presenter/lib/pakyow/presenter/partial.rb
@@ -1,5 +1,6 @@
 module Pakyow
   module Presenter
+    # TODO: refactor to PartialPresenter
     class Partial < View
     end
   end
